Alternating row colors. Widget opens calendar.
Extended simplecursoradapter to merely change alternating background
colors.

Changed the pending intent to habit calendar passing in the habit name and
the widget ID as the "request code" to differentiate the intents.

Also created a new function to remove the key value pairs in share
preferences when the widget is deleted.